Unify coordinate systems
Oof, the player and the obstacles weren't even in the same (X,Y) space.
Now they are.  That lets us drop the temporary obstacleHitBox variable,
which only existed to project obstacles into the player's frame to see if
they overlapped.

All in the name of simplifying the parts the students will be editing.